+++ NewTransport/FileManager/FileManager.Infrastructure/Repositories/TaskRepository.cs
@@ -1,4 +1,5 @@
 ﻿using FileManager.Core.Entities;
+using FileManager.Core.Exceptions;
 using FileManager.Core.Interfaces.Repositories;
 using FileManager.Infrastructure.Data;
 using Microsoft.EntityFrameworkCore;
@@ -11,7 +12,7 @@ public class TaskRepository(AppDbContext dbContext) : ITaskRepository
     {
         task.LastModified = DateTime.Now;
         await dbContext.Task.AddAsync(task);
-        await CreateTaskStatuse(task.TaskId);
+        await CreateTaskStatus(task.TaskId);
         return task;
     }
 
@@ -20,19 +21,17 @@ public class TaskRepository(AppDbContext dbContext) : ITaskRepository
         try
         {
             var task = await GetTaskById(idTask);
+            if (task == null)
+            {
+                return false;
+            }
             dbContext.Task.Remove(task);
-            await dbContext.SaveChangesAsync();
             return true;
         }
         catch (Exception)
         {
             return false;
         }
-    }
-
-    public async Task<List<TaskGroupEntity>> GetAllGroups()
-    {
-        return await dbContext.TaskGroup.ToListAsync();
     }
 
     public async Task<List<TaskEntity>> GetAllTasks()
@@ -40,155 +39,46 @@ public class TaskRepository(AppDbContext dbContext) : ITaskRepository
         return await dbContext.Task.ToListAsync();
     }
 
-    public async Task<TaskEntity> GetTaskById(string idTask)
+    public async Task<TaskEntity?> GetTaskById(string idTask)
     {
         return await dbContext.Task.FirstOrDefaultAsync(x => x.TaskId == idTask);
     }
 
-    public async Task<TaskGroupEntity> GetTaskGroupByName(string groupName)
+    public async Task<TaskEntity> EditTask(TaskEntity task)
     {
-        return await dbContext.TaskGroup.FirstOrDefaultAsync(x => x.Name == groupName);
-    }
-
-    public async Task<List<TaskEntity>> GetTasksByGroup(int idGroup)
-    {
-        return await dbContext.Task.Where(x => x.TaskGroupId == idGroup).ToListAsync();
-    }
-
-    public async Task<bool> EditTask(TaskEntity task)
-    {
-
-        TaskEntity? oldTask = await dbContext.Task.FirstOrDefaultAsync(x => x.TaskId == task.TaskId);
-        if (oldTask != null)
-        {
-            oldTask.Name = task.Name;
-            oldTask.TaskGroupId = task.TaskGroupId;
-            oldTask.AddresseeGroupId = task.AddresseeGroupId;
-            oldTask.IsActive = task.IsActive;
-            oldTask.DayActive = task.DayActive;
-            oldTask.ExecutionLimit = task.ExecutionLimit;
-            oldTask.TimeEnd = task.TimeEnd;
-            oldTask.TimeBegin = task.TimeBegin;
-            oldTask.LastModified = DateTime.Now;
-            dbContext.Task.Update(oldTask);
-            await dbContext.SaveChangesAsync();
-            return true;
-        }
-        return false;
-    }
-
-    public async Task<bool> UpdateLastModifiedTask(string idTask)
-    {
-        try
-        {
-            TaskEntity? task = await dbContext.Task.FirstOrDefaultAsync(x => x.TaskId == idTask);
-            if (task != null)
-            {
-                task.LastModified = DateTime.Now;
-                dbContext.Task.Update(task);
-                await dbContext.SaveChangesAsync();
-                return true;
-            }
-            return false;
-
-        }
-        catch (Exception)
-        {
-            return false;
-        }
-    }
-
-    public async Task<TaskGroupEntity?> CreateTaskGroup(string name)
-    {
-        try
-        {
-            TaskGroupEntity? taskGroups = await dbContext.TaskGroup.FirstOrDefaultAsync(x => x.Name == name);
-            if (taskGroups == null)
-            {
-                TaskGroupEntity taskGroupEntity = new()
-                {
-                    Name = name
-                };
-                await dbContext.TaskGroup.AddAsync(taskGroupEntity);
-                await dbContext.SaveChangesAsync();
-                return taskGroupEntity;
-            }
-            return null;
-        }
-        catch (Exception)
-        {
-            return null;
-        }
-    }
-
-    public async Task<bool> DeleteTaskGroup(int idGroup)
-    {
-        try
-        {
-            List<TaskEntity> tasks = await dbContext.Task.Where(x => x.TaskGroupId == idGroup).ToListAsync();
-            foreach (var task in tasks)
-            {
-                task.TaskGroupId = 0;
-            }
-            await dbContext.SaveChangesAsync();
-            TaskGroupEntity? taskGroup = await dbContext.TaskGroup.FirstOrDefaultAsync(x => x.Id == idGroup);
-            if (taskGroup != null)
-            {
-                dbContext.TaskGroup.Remove(taskGroup);
-                await dbContext.SaveChangesAsync();
-                return true;
-            }
-            return false;
-        }
-        catch (Exception)
-        {
-            return false;
-        }
+        TaskEntity editedTask = await dbContext.Task.FirstOrDefaultAsync(x => x.TaskId == task.TaskId)
+                                        ?? throw new DomainException("Задача с таким Id не найдена");
+        editedTask.Name = task.Name;
+        editedTask.TaskGroupId = task.TaskGroupId;
+        editedTask.AddresseeGroupId = task.AddresseeGroupId;
+        editedTask.IsActive = task.IsActive;
+        editedTask.DayActive = task.DayActive;
+        editedTask.ExecutionLimit = task.ExecutionLimit;
+        editedTask.TimeEnd = task.TimeEnd;
+        editedTask.TimeBegin = task.TimeBegin;
+        editedTask.LastModified = DateTime.Now;
+        dbContext.Task.Update(editedTask);
+        return editedTask;
     }
 
     public async Task<bool> ActivatedTask(string idTask)
     {
-        try
-        {
-            TaskEntity? task = await dbContext.Task.FirstOrDefaultAsync(x => x.TaskId == idTask);
-            if (task != null)
-            {
-                task.IsActive = !task.IsActive;
-                task.LastModified = DateTime.Now;
-                dbContext.Task.Update(task);
-                await dbContext.SaveChangesAsync();
-            }
-            //if (task.IsActive)
-            //{
-            //	_userLogging.Logging(HttpContext.User.Identity.Name, $"Включение задачи: {task.TaskId}", JsonSerializer.Serialize(task));
-            //}
-            //else
-            //{
-            //	_userLogging.Logging(HttpContext.User.Identity.Name, $"Выключение задачи: {task.TaskId}", JsonSerializer.Serialize(task));
-            //}
-            return true;
-        }
-        catch (Exception)
-        {
-            return false;
-        }
+        TaskEntity task = await dbContext.Task.FirstOrDefaultAsync(x => x.TaskId == idTask)
+                                    ?? throw new DomainException("Задача с таким Id не найдена");
+        task.IsActive = !task.IsActive;
+        task.LastModified = DateTime.Now;
+        dbContext.Task.Update(task);
+        return true;
     }
 
-    public async Task<bool> CreateTaskStatuse(string idTask)
+    public async Task<TaskStatusEntity> CreateTaskStatus(string idTask)
     {
-        try
+        TaskStatusEntity taskStatus = new()
         {
-            TaskStatusEntity taskStatuse = new()
-            {
-                TaskId = idTask
-            };
-            await dbContext.TaskStatuse.AddAsync(taskStatuse);
-            return true;
-        }
-        catch (Exception)
-        {
-            return false;
-        }
+            TaskId = idTask
+        };
+        await dbContext.TaskStatuse.AddAsync(taskStatus);
+        return taskStatus;
     }
 
     public async Task<List<TaskStatusEntity>> GetTaskStatuses()
@@ -196,17 +86,9 @@ public class TaskRepository(AppDbContext dbContext) : ITaskRepository
         return await dbContext.TaskStatuse.ToListAsync();
     }
 
-    public async Task<bool> UpdateTaskStatus(TaskStatusEntity taskStatus)
+    public TaskStatusEntity UpdateTaskStatus(TaskStatusEntity taskStatus)
     {
-        try
-        {
-            dbContext.TaskStatuse.Update(taskStatus);
-            await dbContext.SaveChangesAsync();
-            return true;
-        }
-        catch (Exception)
-        {
-            return false;
-        }
+        dbContext.TaskStatuse.Update(taskStatus);
+        return taskStatus;
     }
 }
